More implicit cast fixes. Add in sd test case.
(cherry picked from commit 9aaef89ca6e4db3657178711bceba204d29bbf6b)

+++ libraries/stdlib/genglsl/lib/mx_noise.glsl
@@ -163,8 +163,8 @@ uint mx_hash_int(int x, int y)
     uint a, b, c;
     uint len = 2u;
     a = b = c = uint(0xdeadbeef) + (len << 2u) + 13u;
-    a += x;
-    b += y;
+    a += uint(x);
+    b += uint(y);
     c = mx_bjfinal(a, b, c);
     return c;
 }
@@ -174,9 +174,9 @@ uint mx_hash_int(int x, int y, int z)
     uint a, b, c;
     uint len = 3u;
     a = b = c = uint(0xdeadbeef) + (len << 2u) + 13u;
-    a += x;
-    b += y;
-    c += z;
+    a += uint(x);
+    b += uint(y);
+    c += uint(z);
     c = mx_bjfinal(a, b, c);
     return c;
 }
